requirements.txt
Flask==1.1.2
google-cloud-ndb==1.3.0
google-api-python-client==1.9.3
google-auth-oauthlib==0.4.1
google-auth==1.18.0
google-auth-httplib2==0.0.3
requests==2.24.0